Move dropdown-related stuff to UI component
