+++ app/ui-react/syndesis/src/modules/integrations/components/index.ts
@@ -1,6 +1,5 @@
 export * from './Integrations';
 export * from './IntegrationCreatorBreadcrumbs';
-export * from './IntegrationDetailHistory';
 export * from './IntegrationDetailSteps';
 export * from './IntegrationEditorBreadcrumbs';
 export * from './IntegrationEditorSidebar';
